Count

+++ src/main/java/com/korczak/plsql1/TableCount.java
@@ -22,7 +22,7 @@ public class TableCount extends StoredProcedure {
         compile();
     }
 
-    public BigDecimal execute(String tableName){
+    public Integer execute(String tableName){
         Map<String, String> inputParams = new HashMap();
         inputParams.put("table_name", tableName);
 
@@ -30,9 +30,10 @@ public class TableCount extends StoredProcedure {
         if(!outputParams.isEmpty()) {
             Object rowCount = outputParams.get("table_count");
             if (rowCount instanceof BigDecimal) {
-                return (BigDecimal) rowCount;
+                return (Integer) rowCount;
             }
         }
         return null;
     }
+
 }
